Switched from window to parent element, so list can be embedded
Previously the code relied on the list being a child of the window. Have
switched this so that it just now requires a parent element around the
list, on which scroll events are listened to. It also uses the scroll
properties of this parent element to determine it's position. In future
I'm going to look at wrapping the list in an element, removing the need
for an outer element being manually specified.